Remove static requirement from SystemState params
