+++ crates/bevy_ecs/src/schedule/system_descriptor.rs
@@ -29,7 +29,7 @@ use crate::{
 /// SystemStage::parallel()
 ///     .with_system(do_something.label(Something))
 ///     .with_system(do_the_other_thing.after(Something))
-///     .with_system(do_something_else.exclusive_system().at_end());
+///     .with_system(do_something_else.at_end());
 /// ```
 pub struct SystemDescriptor {
     pub(crate) system: BoxedSystem<(), ()>,
