+++ crates/bevy_render/src/render_phase/draw.rs
@@ -151,7 +151,7 @@ impl<P: PhaseItem> DrawFunctions<P> {
 pub trait RenderCommand<P: PhaseItem> {
     /// Specifies all ECS data required by [`RenderCommand::render`].
     /// All parameters have to be read only.
-    type Param: SystemParam;
+    type Param: SystemParam + 'static;
 
     /// Renders the [`PhaseItem`] by issuing draw calls via the [`TrackedRenderPass`].
     fn render<'w>(
@@ -168,7 +168,7 @@ pub enum RenderCommandResult {
 }
 
 pub trait EntityRenderCommand {
-    type Param: SystemParam;
+    type Param: SystemParam + 'static;
     fn render<'w>(
         view: Entity,
         item: Entity,
@@ -231,7 +231,10 @@ pub enum BatchResult {
     IncompatibleItems,
 }
 
-impl<P: EntityPhaseItem, E: EntityRenderCommand> RenderCommand<P> for E {
+impl<P: EntityPhaseItem, E: EntityRenderCommand> RenderCommand<P> for E
+where
+    E::Param: 'static,
+{
     type Param = E::Param;
 
     #[inline]
@@ -292,7 +295,7 @@ all_tuples!(render_command_tuple_impl, 0, 15, C);
 
 /// Wraps a [`RenderCommand`] into a state so that it can be used as a [`Draw`] function.
 /// Therefore the [`RenderCommand::Param`] is queried from the ECS and passed to the command.
-pub struct RenderCommandState<P: PhaseItem, C: RenderCommand<P>> {
+pub struct RenderCommandState<P: PhaseItem + 'static, C: RenderCommand<P>> {
     state: SystemState<C::Param>,
 }
